requirements.txt
requests
beautifulsoup4
schedule
tensorflow
matplotlib
pillow
numpy
keras
pandas
selenium